general folder refactoring
