+++ examples/Figs/fig1_b.pdf
