+++ examples/Figs/fig1_b.png
