+++ examples/Figs/fig2_b.pdf
