+++ examples/Figs/fig2_b.png
