+++ examples/Figs/fig3_b.pdf
