+++ examples/Figs/fig3_b.png
