+++ examples/Figs/fig3_c.png
