+++ examples/Figs/fig4_b.pdf
