+++ examples/Figs/fig4_b.png
