+++ examples/Figs/fig4_c.png
