+++ pysrc/paramo/paramo.pyi
