+++ requirements.txt
@@ -3,3 +3,4 @@ matplotlib==3.8.3
 numpy==1.26.4
 pytest==8.1.1
 toml==0.10.2
+
